requirements.txt
fastapi
uvicorn

sqlalchemy
psycopg2-binary
pydantic
alembic
python-dotenv
confluent-kafka
requests
httpx
pytest
pytest-cov
